Restructure Parse
