+++ stack.yaml
@@ -1,4 +1,4 @@
-resolver: lts-18.21
+resolver: lts-18.28
 
 packages:
 - .
